requirements.txt
discord.py
anthropic
apscheduler
